Separate lib functions
